Move Headers constant to api
