+++ src/api/clients/IOClients/apps/Builder.ts
@@ -1,6 +1,6 @@
 import { AppClient, CacheType, InstanceOptions, IOContext, RequestConfig } from '@vtex/api'
 import { ChangeToSend } from '../../../modules/apps/ProjectUploader'
-import { Headers } from '../../../../lib/constants/Headers'
+import { Headers } from '../../../constants/Headers'
 import { ErrorKinds } from '../../../error/ErrorKinds'
 import { ErrorReport } from '../../../error/ErrorReport'
 import { IOClientFactory } from '../IOClientFactory'
