+++ src/api/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './Colors'
 export * from './Messages'
+export * from './Headers'
