+++ src/lib/sse/CustomEventSource.ts
@@ -1,7 +1,7 @@
 import EventSource from 'eventsource'
 import { cluster, envCookies } from '../../api/env'
 import userAgent from '../../user-agent'
-import { Headers } from '../constants/Headers'
+import { Headers } from '../../api/constants/Headers'
 import { SessionManager } from '../../api/session/SessionManager'
 import { EventSourceError } from './EventSourceError'
 import { TraceConfig } from '../globalConfigs/traceConfig'
