+++ src/modules/apps/debugger.ts
@@ -2,7 +2,7 @@ import streamToString from 'get-stream'
 import net from 'net'
 import WebSocket from 'ws'
 import { cluster } from '../../api/env'
-import { Headers } from '../../lib/constants/Headers'
+import { Headers } from '../../api/constants/Headers'
 import { ManifestEditor } from '../../api/manifest'
 import { SessionManager } from '../../api/session/SessionManager'
 import { versionMajor } from '../../api/locator'
